Rename: user 패키지명 member로 수정
