+++ src/main/java/com/zerobase/CafeBom/member/domain/entity/Member.java
@@ -1,8 +1,8 @@
-package com.zerobase.cafebom.user.domain.entity;
+package com.zerobase.cafebom.member.domain.entity;
 
 import com.zerobase.cafebom.common.BaseTimeEntity;
-import com.zerobase.cafebom.user.security.Role;
-import com.zerobase.cafebom.user.service.dto.SignupDto;
+import com.zerobase.cafebom.member.security.Role;
+import com.zerobase.cafebom.member.service.dto.SignupDto;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
