+++ src/main/java/com/zerobase/CafeBom/member/security/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.zerobase.cafebom.user.security;
+package com.zerobase.cafebom.member.security;
 
 import java.io.IOException;
 import javax.servlet.FilterChain;
