+++ src/main/java/com/zerobase/CafeBom/member/security/Role.java
@@ -1,4 +1,4 @@
-package com.zerobase.cafebom.user.security;
+package com.zerobase.cafebom.member.security;
 
 public enum Role {
     ROLE_USER,
