+++ src/main/java/com/zerobase/CafeBom/member/security/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.zerobase.cafebom.user.security;
+package com.zerobase.cafebom.member.security;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
