+++ src/main/java/com/zerobase/CafeBom/member/security/TokenProvider.java
@@ -1,7 +1,7 @@
-package com.zerobase.cafebom.user.security;
+package com.zerobase.cafebom.member.security;
 
 
-import com.zerobase.cafebom.user.service.AuthService;
+import com.zerobase.cafebom.member.service.AuthService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
