+++ src/main/java/com/zerobase/cafebom/orders/domain/entity/Order.java
@@ -3,7 +3,7 @@ package com.zerobase.cafebom.orders.domain.entity;
 import com.zerobase.cafebom.common.BaseTimeEntity;
 import com.zerobase.cafebom.orders.domain.type.OrdersCookingTime;
 import com.zerobase.cafebom.orders.domain.type.OrdersStatus;
-import com.zerobase.cafebom.user.domain.entity.Member;
+import com.zerobase.cafebom.member.domain.entity.Member;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
