+++ src/test/java/com/zerobase/CafeBom/member/repository/MemberRepositoryTest.java
@@ -1,5 +1,6 @@
-package com.zerobase.cafebom.user.repository;
+package com.zerobase.CafeBom.member.repository;
 
+import com.zerobase.cafebom.member.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
